refactor: reorganize into separate js files
